(WIP) Converted Body to struct
This is a radical change to our API, and the infrastructure required to make this work efficiently is rather complicated. I'm going to explore the possibility of making Fixture and Joint structs (I'm fairly confident this will work nicely with Fixtures but Joints could be a bit trickier due to inheritance).